% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.mpear.R
\name{adjust.mpear}
\alias{adjust.mpear}
\title{Adjust MPEAR}
\usage{
adjust.mpear(psm, method = "ward.D2", max.k = NULL)
}
\arguments{
\item{psm}{is posterior similarity matrix with entries between 0 and 1 and 1's on the diagonals which is distance matrix.}

\item{method}{is clustering method. We offer many kinds of method, like as kmeans and hierarchical clustering. The default is ward.D2 of hierarchical clustering.}

\item{max.k}{is number with the user has the professional judgment to consider the max group and the default is NULL.}
}
\value{
cl is vector with the clustering result.

value is number with MPEAR.

method is clustering method.
}
\description{
adjust.mpear is modified the mcclust package to suit our model.
}
\details{
adjust.mpear is modified the mcclust package to suit our model. And the function use distance matrix by criteria of MPEAR to cluster the data.
}
\author{
Peter Wu (peter123wu0@gmail.com)
}
\references{
Fritsch, A. and Ickstadt, K. Improved criteria for clustering based on the posterior similarity matrix. Bayesian analysis 2009;4(2):367-391.
}

