% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hamming.R
\name{Hamming}
\alias{Hamming}
\title{Hamming distance}
\usage{
Hamming(k.matrix)
}
\arguments{
\item{k.matrix}{is matrix with the result-retaining of MCMC.}
}
\value{
Hamming.matrix is the co-occurrence matrix and the idea is Hamming distance.

pro.matrix is the standardization of co-occurrence matrix.

data is data.frame with matrix. The purpose is constructing the heat map.
}
\description{
Hamming is computing the co-occurrence matrix.
}
\details{
Hamming use the result-retaining of MCMC to compute the co-occurrence matrix. And we use the co-occurrence matrix to cluster which the criteria is MPEAR.
}
\author{
Peter Wu (peter123wu0@gmail.com)
}

