% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_similar.R
\name{tf_similar}
\alias{tf_similar}
\title{True or false of similar}
\usage{
tf_similar(real, cluster)
}
\arguments{
\item{real}{is vector with the real group number. If we don't have the real group number, we also can input the result of some clustering method.}

\item{cluster}{is vector with the result of clustering.}
}
\value{
result is data.frame with 10 columns which contains the TT, TF, FT, FF, rand index(RI), weight rand index(WRI), adjust rand index(ARI), the weighted TT, the weighted TF, and the weight FT.

weight is data.frame with 3 columns which contains the weight about the TT, TF, and FT.
}
\description{
tf_similar is comparing the similarity of different group.
}
\details{
tf_similar is comparing the similarity of different group. So the main purpose is checking the quality of clustering method. We show the rand index(RI), weight rand index(WRI), and adjust rand index(ARI) to compare. In order to compute the three kinds of index, we count the TT, TF, FT, FF, and the relation weight. The details of the function of the function  are in paper of Tumor subclones detection with Dirichlet Process Mixture Model.
}
\author{
Peter Wu (peter123wu0@gmail.com)
}

